chenged access of getting a portfolio and accessing blog posts
Now anyone can access blog posts and portfolios, privacy should be something set within a portfolio. Later in these calls there should be a check to see if the portfolio/blog is private or public, and give an error depending on that. But anyone should be able to make the calls to these functions without a token